pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iptax"
version = "0.1.0"
description = "Automated IP tax report generator for Polish tax deduction program"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Chris Suszyński", email = "krzysztof.suszynski@gmail.com"}
]
keywords = ["tax", "reporting", "ip-tax", "automation", "polish-tax"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    # CLI framework
    "click>=8.1.7",

    # Configuration management
    "pyyaml>=6.0.1",
    "tomli-w>=1.0.0",  # TOML writing (reading via stdlib tomllib)

    # AI provider integration
    "litellm>=1.44.0",
    "google-generativeai>=0.3.0",

    # Date handling
    "python-dateutil>=2.8.2",

    # Browser automation for Workday
    "playwright>=1.40.0",

    # PDF generation
    "weasyprint>=60.0",
    "jinja2>=3.1.2",

    # HTTP client
    "httpx>=0.25.0",

    # Markdown processing
    "markdown>=3.5.0",

    # Rich terminal UI
    "rich>=13.7.0",
    "questionary>=2.0.1",

    # Data validation
    "pydantic>=2.5.0",

    # Did integration (PR/MR fetching) - using PR #311
    "did @ git+https://github.com/psss/did.git@refs/pull/311/head",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-playwright>=0.4.3",
    "pytest-asyncio>=0.21.0",
    "pytest-timeout>=2.2.0",  # For test timeouts
    "pytest-xdist>=3.5.0",  # For parallel test execution
    "flask>=3.0.0",  # For e2e mock servers
    "werkzeug>=3.0.0",  # Flask dependency for e2e mock servers

    # Code quality
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "editorconfig-checker>=2.7.3",
    "mdformat>=0.7.17",
    "mdformat-gfm>=0.3.5",  # GitHub Flavored Markdown support
    "mdformat-frontmatter>=2.0.8",  # YAML frontmatter support

    # Type stubs
    "types-pyyaml>=6.0.12",
    "types-python-dateutil>=2.8.19",

    # Development tools
    "ipython>=8.18.0",
    "ipdb>=0.13.13",
]

[project.scripts]
iptax = "iptax.cli:main"

[project.urls]
Homepage = "https://github.com/cardil/iptax"
Repository = "https://github.com/cardil/iptax"
Issues = "https://github.com/cardil/iptax/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["iptax*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "TRY",  # tryceratops - exception handling
    "S",    # flake8-bandit - security
    "PL",   # pylint
    "RUF",  # ruff-specific rules
    "ANN",  # flake8-annotations
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "PTH",  # flake8-use-pathlib
    "T20",  # flake8-print
    "Q",    # flake8-quotes
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "ICN",  # flake8-import-conventions
]
ignore = [
    "E501",    # line too long, handled by black
    "B008",    # do not perform function calls in argument defaults
    "S101",    # use of assert detected (pytest uses assert)
    "TRY003",  # avoid specifying long messages outside exception class
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = [
    "ANN001",  # missing type annotations for function arguments (pytest fixtures)
    "ANN201",  # missing return type annotations (tests return None implicitly)
    "PLR2004", # magic values (common in test assertions)
    "PLC0415", # import at top-level (lazy imports in tests are acceptable)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
plugins = ['pydantic.mypy']

[[tool.mypy.overrides]]
module = [
    "playwright.*",
    "weasyprint.*",
    "litellm.*",
    "did.*",
    "tomli_w",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=iptax",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
    "requires_did: Tests requiring did installation",
    "requires_workday: Tests requiring Workday access",
]
# Use thread-based timeout to allow proper cleanup of daemon threads
timeout_method = "thread"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
